[#106] rename: placeCard2로 -> placeCard2로 수정
