+++ src/components/tourism/PlaceCard2.tsx
